# openapi_server/models/create_review.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class CreateReview(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    CreateReview - a model defined in OpenAPI

        user_id: The user_id of this CreateReview.
        restaurant_id: The restaurant_id of this CreateReview.
        rating: The rating of this CreateReview.
        content: The content of this CreateReview [Optional].
        photo_url: The photo_url of this CreateReview [Optional].
    """

    user_id: int = Field(alias="userId")
    restaurant_id: int = Field(alias="restaurantId")
    rating: int = Field(alias="rating")
    content: Optional[str] = Field(alias="content", default=None)
    photo_url: Optional[str] = Field(alias="photoUrl", default=None)


CreateReview.update_forward_refs()