# openapi_server/models/favorite_food.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.review import Review


class FavoriteFood(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    FavoriteFood - a model defined in OpenAPI

        id: The id of this FavoriteFood.
        name: The name of this FavoriteFood.
        reviews: The reviews of this FavoriteFood.
    """

    id: int = Field(alias="id")
    name: str = Field(alias="name")
    reviews: List[Review] = Field(alias="reviews")


FavoriteFood.update_forward_refs()