# openapi_server/models/list_reviews.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.review import Review


class ListReviews(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    ListReviews - a model defined in OpenAPI

        reviews: The reviews of this ListReviews.
    """

    reviews: List[Review] = Field(alias="reviews")


ListReviews.update_forward_refs()