# openapi_server/models/login_user.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class LoginUser(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    LoginUser - a model defined in OpenAPI

        username_or_email: The username_or_email of this LoginUser.
        password: The password of this LoginUser.
    """

    username_or_email: str = Field(alias="usernameOrEmail")
    password: str = Field(alias="password")


LoginUser.update_forward_refs()