# openapi_server/models/update_review.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class UpdateReview(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    UpdateReview - a model defined in OpenAPI

        rating: The rating of this UpdateReview [Optional].
        content: The content of this UpdateReview [Optional].
        photo_url: The photo_url of this UpdateReview [Optional].
    """

    rating: Optional[int] = Field(alias="rating", default=None)
    content: Optional[str] = Field(alias="content", default=None)
    photo_url: Optional[str] = Field(alias="photoUrl", default=None)


UpdateReview.update_forward_refs()