# openapi_server/models/user.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.favorite_food import FavoriteFood


class User(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    User - a model defined in OpenAPI

        id: The id of this User.
        username: The username of this User.
        first_name: The first_name of this User.
        last_name: The last_name of this User.
        email: The email of this User.
        password: The password of this User.
        favorite_foods: The favorite_foods of this User.
        friends: The friends of this User.
    """

    id: int = Field(alias="id")
    username: str = Field(alias="username")
    first_name: str = Field(alias="firstName")
    last_name: str = Field(alias="lastName")
    email: str = Field(alias="email")
    password: str = Field(alias="password")
    favorite_foods: List[FavoriteFood] = Field(alias="favoriteFoods")
    friends: List[User] = Field(alias="friends")


User.update_forward_refs()